Removed leftover exercises
