--- workshops/introduction_to_event_sourcing/src/06_business_logic_eventstoredb/oop/businessLogic.exercise.test.ts
@@ -1,316 +1,0 @@
-import { getEventStoreDBTestClient } from '#core/testing/eventStoreDB';
-import { EventStoreDBClient } from '@eventstore/db-client';
-import { v4 as uuid } from 'uuid';
-
-export interface ProductItem {
-  productId: string;
-  quantity: number;
-}
-
-export type PricedProductItem = ProductItem & {
-  unitPrice: number;
-};
-
-export type Event<
-  EventType extends string = string,
-  EventData extends Record<string, unknown> = Record<string, unknown>,
-> = Readonly<{
-  type: Readonly<EventType>;
-  data: Readonly<EventData>;
-}>;
-
-export type ShoppingCartOpened = Event<
-  'ShoppingCartOpened',
-  {
-    shoppingCartId: string;
-    clientId: string;
-    openedAt: string;
-  }
->;
-
-export type ProductItemAddedToShoppingCart = Event<
-  'ProductItemAddedToShoppingCart',
-  {
-    shoppingCartId: string;
-    productItem: PricedProductItem;
-  }
->;
-
-export type ProductItemRemovedFromShoppingCart = Event<
-  'ProductItemRemovedFromShoppingCart',
-  {
-    shoppingCartId: string;
-    productItem: PricedProductItem;
-  }
->;
-
-export type ShoppingCartConfirmed = Event<
-  'ShoppingCartConfirmed',
-  {
-    shoppingCartId: string;
-    confirmedAt: string;
-  }
->;
-
-export type ShoppingCartCanceled = Event<
-  'ShoppingCartCanceled',
-  {
-    shoppingCartId: string;
-    canceledAt: string;
-  }
->;
-
-export type ShoppingCartEvent =
-  | ShoppingCartOpened
-  | ProductItemAddedToShoppingCart
-  | ProductItemRemovedFromShoppingCart
-  | ShoppingCartConfirmed
-  | ShoppingCartCanceled;
-
-export enum ShoppingCartStatus {
-  Pending = 'Pending',
-  Confirmed = 'Confirmed',
-  Canceled = 'Canceled',
-}
-
-export class ShoppingCart {
-  constructor(
-    private _id: string,
-    private _clientId: string,
-    private _status: ShoppingCartStatus,
-    private _openedAt: Date,
-    private _productItems: PricedProductItem[] = [],
-    private _confirmedAt?: Date,
-    private _canceledAt?: Date,
-  ) {}
-
-  get id() {
-    return this._id;
-  }
-
-  get clientId() {
-    return this._clientId;
-  }
-
-  get status() {
-    return this._status;
-  }
-
-  get openedAt() {
-    return this._openedAt;
-  }
-
-  get productItems() {
-    return this._productItems;
-  }
-
-  get confirmedAt() {
-    return this._confirmedAt;
-  }
-
-  get canceledAt() {
-    return this._canceledAt;
-  }
-
-  public evolve = ({ type, data: event }: ShoppingCartEvent): void => {
-    switch (type) {
-      case 'ShoppingCartOpened': {
-        this._id = event.shoppingCartId;
-        this._clientId = event.clientId;
-        this._status = ShoppingCartStatus.Pending;
-        this._openedAt = new Date(event.openedAt);
-        this._productItems = [];
-        return;
-      }
-      case 'ProductItemAddedToShoppingCart': {
-        const {
-          productItem: { productId, quantity, unitPrice },
-        } = event;
-
-        const currentProductItem = this._productItems.find(
-          (pi) => pi.productId === productId && pi.unitPrice === unitPrice,
-        );
-
-        if (currentProductItem) {
-          currentProductItem.quantity += quantity;
-        } else {
-          this._productItems.push(event.productItem);
-        }
-        return;
-      }
-      case 'ProductItemRemovedFromShoppingCart': {
-        const {
-          productItem: { productId, quantity, unitPrice },
-        } = event;
-
-        const currentProductItem = this._productItems.find(
-          (pi) => pi.productId === productId && pi.unitPrice === unitPrice,
-        );
-
-        if (!currentProductItem) {
-          return;
-        }
-
-        currentProductItem.quantity -= quantity;
-
-        if (currentProductItem.quantity <= 0) {
-          this._productItems.splice(
-            this._productItems.indexOf(currentProductItem),
-            1,
-          );
-        }
-        return;
-      }
-      case 'ShoppingCartConfirmed': {
-        this._status = ShoppingCartStatus.Confirmed;
-        this._confirmedAt = new Date(event.confirmedAt);
-        return;
-      }
-      case 'ShoppingCartCanceled': {
-        this._status = ShoppingCartStatus.Canceled;
-        this._canceledAt = new Date(event.canceledAt);
-        return;
-      }
-    }
-  };
-}
-
-export const getShoppingCart = (events: ShoppingCartEvent[]): ShoppingCart => {
-  return events.reduce<ShoppingCart>(
-    (state, event) => {
-      state.evolve(event);
-      return state;
-    },
-    new ShoppingCart(
-      undefined!,
-      undefined!,
-      undefined!,
-      undefined!,
-      undefined,
-      undefined,
-      undefined,
-    ),
-  );
-};
-
-export const mapShoppingCartStreamId = (id: string) => `shopping_cart-${id}`;
-
-export const readStream = async (
-  eventStore: EventStoreDBClient,
-  shoppingCartId: string,
-) => {
-  const readResult = eventStore.readStream<ShoppingCartEvent>(
-    mapShoppingCartStreamId(shoppingCartId),
-  );
-
-  const events: ShoppingCartEvent[] = [];
-
-  for await (const { event } of readResult) {
-    if (!event) continue;
-    events.push(<ShoppingCartEvent>{ type: event.type, data: event.data });
-  }
-
-  return events;
-};
-
-describe('Getting state from events', () => {
-  let eventStore: EventStoreDBClient;
-
-  beforeAll(async () => {
-    eventStore = await getEventStoreDBTestClient();
-  });
-
-  it('Should return the state from the sequence of events', async () => {
-    const shoppingCartId = uuid();
-
-    const clientId = uuid();
-    const openedAt = new Date();
-    const confirmedAt = new Date();
-    // const canceledAt = new Date();
-
-    const shoesId = uuid();
-
-    const twoPairsOfShoes: PricedProductItem = {
-      productId: shoesId,
-      quantity: 2,
-      unitPrice: 100,
-    };
-    const pairOfShoes: PricedProductItem = {
-      productId: shoesId,
-      quantity: 1,
-      unitPrice: 100,
-    };
-
-    const tShirtId = uuid();
-    const tShirt: PricedProductItem = {
-      productId: tShirtId,
-      quantity: 1,
-      unitPrice: 5,
-    };
-
-    // TODO: Fill the events store results of your business logic
-    // to be the same as events below
-
-    const events = await readStream(eventStore, shoppingCartId);
-
-    expect(events).toEqual([
-      {
-        type: 'ShoppingCartOpened',
-        data: {
-          shoppingCartId,
-          clientId,
-          openedAt: openedAt.toISOString(),
-        },
-      },
-      {
-        type: 'ProductItemAddedToShoppingCart',
-        data: {
-          shoppingCartId,
-          productItem: twoPairsOfShoes,
-        },
-      },
-      {
-        type: 'ProductItemAddedToShoppingCart',
-        data: {
-          shoppingCartId,
-          productItem: tShirt,
-        },
-      },
-      {
-        type: 'ProductItemRemovedFromShoppingCart',
-        data: { shoppingCartId, productItem: pairOfShoes },
-      },
-      {
-        type: 'ShoppingCartConfirmed',
-        data: {
-          shoppingCartId,
-          confirmedAt: confirmedAt.toISOString(),
-        },
-      },
-      // This should fail
-      // {
-      //   type: 'ShoppingCartCanceled',
-      //   data: {
-      //     shoppingCartId,
-      //     canceledAt: canceledAt.toISOString(),
-      //   },
-      // },
-    ]);
-
-    const shoppingCart = getShoppingCart(events);
-
-    expect(shoppingCart).toBeInstanceOf(ShoppingCart);
-    expect(JSON.stringify(shoppingCart)).toBe(
-      JSON.stringify(
-        new ShoppingCart(
-          shoppingCartId,
-          clientId,
-          ShoppingCartStatus.Confirmed,
-          openedAt,
-          [pairOfShoes, tShirt],
-          confirmedAt,
-        ),
-      ),
-    );
-  });
-});
