+++ workshops/introduction_to_event_sourcing/src/solved/06_application_logic_eventstoredb/oop/aggregate_returning_events/api.ts
@@ -6,6 +6,7 @@ import {
 import { sendCreated } from '../../tools/api';
 import { v4 as uuid } from 'uuid';
 import { PricedProductItem, ProductItem } from './shoppingCart';
+import { ShoppingCartService } from './applicationService';
 
 export const mapShoppingCartStreamId = (id: string) => `shopping_cart-${id}`;
 
@@ -13,89 +14,114 @@ const dummyPriceProvider = (_productId: string) => {
   return 100;
 };
 
-export const shoppingCartApi = (router: Router) => {
-  // Open Shopping cart
-  router.post(
-    '/clients/:clientId/shopping-carts/',
-    (request: Request, response: Response) => {
-      const shoppingCartId = uuid();
-      const _clientId = assertNotEmptyString(request.params.clientId);
+export const shoppingCartApi =
+  (shoppingCartService: ShoppingCartService) => (router: Router) => {
+    // Open Shopping cart
+    router.post(
+      '/clients/:clientId/shopping-carts/',
+      async (request: Request, response: Response) => {
+        const shoppingCartId = uuid();
+        const clientId = assertNotEmptyString(request.params.clientId);
 
-      // Fill the gap here
-      throw new Error('Not Implemented!');
+        await shoppingCartService.open({
+          type: 'OpenShoppingCart',
+          data: {
+            shoppingCartId,
+            clientId,
+            now: new Date(),
+          },
+        });
 
-      sendCreated(response, shoppingCartId);
-    },
-  );
+        sendCreated(response, shoppingCartId);
+      },
+    );
 
-  router.post(
-    '/clients/:clientId/shopping-carts/:shoppingCartId/product-items',
-    (request: AddProductItemRequest, response: Response) => {
-      const _shoppingCartId = assertNotEmptyString(
-        request.params.shoppingCartId,
-      );
-      const _productItem: ProductItem = {
-        productId: assertNotEmptyString(request.body.productId),
-        quantity: assertPositiveNumber(request.body.quantity),
-      };
+    router.post(
+      '/clients/:clientId/shopping-carts/:shoppingCartId/product-items',
+      async (request: AddProductItemRequest, response: Response) => {
+        const shoppingCartId = assertNotEmptyString(
+          request.params.shoppingCartId,
+        );
+        const productItem: ProductItem = {
+          productId: assertNotEmptyString(request.body.productId),
+          quantity: assertPositiveNumber(request.body.quantity),
+        };
+        const unitPrice = dummyPriceProvider(productItem.productId);
 
-      // Fill the gap here
-      throw new Error('Not Implemented!');
+        await shoppingCartService.addProductItem({
+          type: 'AddProductItemToShoppingCart',
+          data: {
+            shoppingCartId,
+            productItem: {
+              ...productItem,
+              unitPrice,
+            },
+          },
+        });
 
-      response.sendStatus(204);
-    },
-  );
+        response.sendStatus(204);
+      },
+    );
 
-  // Remove Product Item
-  router.delete(
-    '/clients/:clientId/shopping-carts/:shoppingCartId/product-items',
-    (request: Request, response: Response) => {
-      const _shoppingCartId = assertNotEmptyString(
-        request.params.shoppingCartId,
-      );
-      const _productItem: PricedProductItem = {
-        productId: assertNotEmptyString(request.query.productId),
-        quantity: assertPositiveNumber(Number(request.query.quantity)),
-        unitPrice: assertPositiveNumber(Number(request.query.unitPrice)),
-      };
+    // Remove Product Item
+    router.delete(
+      '/clients/:clientId/shopping-carts/:shoppingCartId/product-items',
+      async (request: Request, response: Response) => {
+        const shoppingCartId = assertNotEmptyString(
+          request.params.shoppingCartId,
+        );
+        const productItem: PricedProductItem = {
+          productId: assertNotEmptyString(request.query.productId),
+          quantity: assertPositiveNumber(Number(request.query.quantity)),
+          unitPrice: assertPositiveNumber(Number(request.query.unitPrice)),
+        };
 
-      // Fill the gap here
-      throw new Error('Not Implemented!');
+        await shoppingCartService.removeProductItem({
+          type: 'RemoveProductItemFromShoppingCart',
+          data: {
+            shoppingCartId,
+            productItem,
+          },
+        });
 
-      response.sendStatus(204);
-    },
-  );
+        response.sendStatus(204);
+      },
+    );
 
-  // Confirm Shopping Cart
-  router.post(
-    '/clients/:clientId/shopping-carts/:shoppingCartId/confirm',
-    (request: Request, response: Response) => {
-      const _shoppingCartId = assertNotEmptyString(
-        request.params.shoppingCartId,
-      );
+    // Confirm Shopping Cart
+    router.post(
+      '/clients/:clientId/shopping-carts/:shoppingCartId/confirm',
+      async (request: Request, response: Response) => {
+        const shoppingCartId = assertNotEmptyString(
+          request.params.shoppingCartId,
+        );
 
-      // Fill the gap here
-      throw new Error('Not Implemented!');
+        await shoppingCartService.confirm({
+          type: 'ConfirmShoppingCart',
+          data: { shoppingCartId, now: new Date() },
+        });
 
-      response.sendStatus(204);
-    },
-  );
+        response.sendStatus(204);
+      },
+    );
 
-  // Cancel Shopping Cart
-  router.delete(
-    '/clients/:clientId/shopping-carts/:shoppingCartId',
-    (request: Request, response: Response) => {
-      const _shoppingCartId = assertNotEmptyString(
-        request.params.shoppingCartId,
-      );
+    // Cancel Shopping Cart
+    router.delete(
+      '/clients/:clientId/shopping-carts/:shoppingCartId',
+      async (request: Request, response: Response) => {
+        const shoppingCartId = assertNotEmptyString(
+          request.params.shoppingCartId,
+        );
 
-      // Fill the gap here
-      throw new Error('Not Implemented!');
+        await shoppingCartService.cancel({
+          type: 'CancelShoppingCart',
+          data: { shoppingCartId, now: new Date() },
+        });
 
-      response.sendStatus(204);
-    },
-  );
-};
+        response.sendStatus(204);
+      },
+    );
+  };
 
 // Add Product Item
 type AddProductItemRequest = Request<
